+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Fixed
+
+- FTP: retry connection on `SSHException` while opening a remote url.
+
 ## [0.12.0] - 2023-09-01
 
 ### Changed